refactor game lobby
